Move participants & sharing to separate components
