+++ plugins/love-resources/src/components/ParticipantView.svelte
@@ -76,7 +76,6 @@
   :global(.video) {
     object-fit: cover;
     border-radius: 0.75rem;
-    height: 100%;
     width: 100%;
   }
   :global(.hidden) {
