+++ plugins/love-resources/src/components/Room.svelte
@@ -13,212 +13,32 @@
 // limitations under the License.
 -->
 <script lang="ts">
-  import { Analytics } from '@hcengineering/analytics'
   import presentation, { ActionContext } from '@hcengineering/presentation'
-  import { getPersonRefByPersonIdCb } from '@hcengineering/contact-resources'
   import { Room as TypeRoom } from '@hcengineering/love'
   import { getMetadata } from '@hcengineering/platform'
-  import { Label, Loading, resizeObserver, deviceOptionsStore as deviceInfo } from '@hcengineering/ui'
-  import {
-    LocalParticipant,
-    LocalTrackPublication,
-    Participant,
-    RemoteParticipant,
-    RemoteTrack,
-    RemoteTrackPublication,
-    RoomEvent,
-    Track,
-    TrackPublication
-  } from 'livekit-client'
-  import { onDestroy, onMount, tick } from 'svelte'
-  import { aiBotSocialIdentityStore } from '@hcengineering/ai-bot-resources'
+  import { Label, Loading, deviceOptionsStore as deviceInfo } from '@hcengineering/ui'
+  import { onDestroy, onMount } from 'svelte'
 
   import love from '../plugin'
   import { waitForOfficeLoaded, currentRoom, infos, invites, myInfo, myRequests } from '../stores'
   import {
-    isFullScreen, liveKitClient,
-    lk,
+    isFullScreen,
     screenSharing,
     tryConnect
   } from '../utils'
   import ControlBar from './ControlBar.svelte'
-  import ParticipantView from './ParticipantView.svelte'
-  import { Ref } from '@hcengineering/core'
-  import { Person } from '@hcengineering/contact'
-  import { ParticipantData } from '../types'
   import { lkSessionConnected } from '../liveKitClient'
+  import ParticipantsListView from './meeting/ParticipantsListView.svelte'
+  import ScreenSharingView from './meeting/ScreenSharingView.svelte'
 
   export let withVideo: boolean
   export let canMaximize: boolean = true
   export let room: TypeRoom
 
-  let participants: ParticipantData[] = []
-  const participantElements: ParticipantView[] = []
-  let screen: HTMLVideoElement
   let roomEl: HTMLDivElement
-
-  let aiPersonRef: Ref<Person> | undefined
-  $: if ($aiBotSocialIdentityStore != null) {
-    getPersonRefByPersonIdCb($aiBotSocialIdentityStore?._id, (ref) => {
-      if (ref != null) {
-        aiPersonRef = ref
-      }
-    })
-  } else {
-    aiPersonRef = undefined
-  }
-
-  function handleTrackSubscribed (
-    track: RemoteTrack,
-    publication: RemoteTrackPublication,
-    participant: RemoteParticipant
-  ): void {
-    if (track.kind === Track.Kind.Video) {
-      if (track.source === Track.Source.ScreenShare) {
-        track.attach(screen)
-      } else {
-        const element = track.attach()
-        attachTrack(element, participant)
-      }
-      updateStyle(getActiveParticipants(participants).length, $screenSharing)
-    } else {
-      const part = participants.find((p) => p._id === participant.identity)
-      if (part !== undefined) {
-        part.muted = publication.isMuted
-        participants = participants
-      }
-    }
-  }
-
-  function handleTrackUnsubscribed (
-    track: RemoteTrack,
-    publication: RemoteTrackPublication,
-    participant: RemoteParticipant
-  ): void {
-    if (track.kind === Track.Kind.Video) {
-      if (track.source !== Track.Source.ScreenShare) {
-        const index = participants.findIndex((p) => p._id === participant.identity)
-        if (index !== -1) {
-          participants.splice(index, 1)
-          participants = participants
-        }
-      } else {
-        track.detach(screen)
-      }
-      updateStyle(getActiveParticipants(participants).length, $screenSharing)
-    }
-  }
-
-  function handleLocalTrack (publication: LocalTrackPublication, participant: LocalParticipant): void {
-    if (publication.track?.kind === Track.Kind.Video) {
-      if (publication.track.source === Track.Source.ScreenShare) {
-        publication.track.attach(screen)
-      } else {
-        const element = publication.track.attach()
-        void attachTrack(element, participant)
-      }
-      updateStyle(getActiveParticipants(participants).length, $screenSharing)
-    } else {
-      const part = participants.find((p) => p._id === participant.identity)
-      if (part !== undefined) {
-        part.muted = publication.isMuted
-        participants = participants
-      }
-    }
-  }
-
-  async function attachTrack (element: HTMLMediaElement, participant: Participant): Promise<void> {
-    let index = participants.findIndex((p) => p._id === participant.identity)
-    if (index === -1) {
-      index = participants.push({
-        _id: participant.identity,
-        name: participant.name ?? '',
-        muted: !participant.isMicrophoneEnabled,
-        mirror: participant.isLocal,
-        connecting: false,
-        isAgent: participant.isAgent
-      })
-    }
-    participants = participants
-    for (let attempt = 0; attempt < 10; attempt++) {
-      await tick()
-      index = participants.findIndex((p) => p._id === participant.identity)
-      const el = participantElements[index]
-      if (el != null) {
-        el.appendChild(element)
-        return
-      }
-    }
-    console.error('Failed to attach track after 10 attempts')
-    Analytics.handleError(new Error(`Failed to attach track after 10 attempts, participant: ${participant.identity}`))
-  }
-
-  function attachParticipant (participant: Participant): void {
-    const current = participants.find((p) => p._id === participant.identity)
-    if (current !== undefined) {
-      current.connecting = false
-      current.muted = !participant.isMicrophoneEnabled
-      current.mirror = participant.isLocal
-      participants = participants
-      return
-    }
-    const value: ParticipantData = {
-      _id: participant.identity,
-      name: participant.name ?? '',
-      muted: !participant.isMicrophoneEnabled,
-      mirror: participant.isLocal,
-      connecting: false,
-      isAgent: participant.isAgent
-    }
-    participants.push(value)
-    participants = participants
-    updateStyle(getActiveParticipants(participants).length, $screenSharing)
-  }
-
-  function handleParticipantDisconnected (participant: RemoteParticipant): void {
-    const index = participants.findIndex((p) => p._id === participant.identity)
-    if (index !== -1) {
-      participants.splice(index, 1)
-      participants = participants
-    }
-    updateStyle(getActiveParticipants(participants).length, $screenSharing)
-  }
-
-  function muteHandler (publication: TrackPublication, participant: Participant): void {
-    if (publication.kind === Track.Kind.Video) {
-      if (publication.source === Track.Source.ScreenShare) {
-        return
-      }
-      const index = participants.findIndex((p) => p._id === participant.identity)
-      if (index !== -1 && participantElements[index] != null) {
-        participantElements[index].setTrackMuted(publication.isMuted)
-      }
-    } else {
-      const part = participants.find((p) => p._id === participant.identity)
-      if (part !== undefined) {
-        part.muted = publication.isMuted
-        participants = participants
-      }
-    }
-  }
 
   let loading: boolean = false
   let configured: boolean = false
-
-  function handleLocalTrackUnsubscribed (publication: LocalTrackPublication, participant: LocalParticipant): void {
-    if (publication?.track?.kind === Track.Kind.Video) {
-      if (publication.track.source === Track.Source.ScreenShare) {
-        publication.track.detach(screen)
-        updateStyle(getActiveParticipants(participants).length, $screenSharing)
-      } else {
-        const index = participants.findIndex((p) => p._id === participant.identity)
-        if (index !== -1) {
-          participants.splice(index, 1)
-          participants = participants
-        }
-      }
-    }
-  }
 
   onMount(async () => {
     loading = true
@@ -240,40 +60,6 @@
       const info = $infos.filter((p) => p.room === room._id)
       await tryConnect($myInfo, room, info, $myRequests, $invites)
     }
-
-    await liveKitClient.awaitConnect()
-    for (const participant of lk.remoteParticipants.values()) {
-      attachParticipant(participant)
-      for (const publication of participant.trackPublications.values()) {
-        if (publication.track !== undefined && publication.track.kind === Track.Kind.Video) {
-          if (publication.track.source === Track.Source.ScreenShare) {
-            publication.track.attach(screen)
-          } else {
-            const element = publication.track.attach()
-            await attachTrack(element, participant)
-          }
-        }
-      }
-    }
-    attachParticipant(lk.localParticipant)
-    for (const publication of lk.localParticipant.trackPublications.values()) {
-      if (publication.track !== undefined && publication.track.kind === Track.Kind.Video) {
-        if (publication.track.source === Track.Source.ScreenShare) {
-          publication.track.attach(screen)
-        } else {
-          const element = publication.track.attach()
-          await attachTrack(element, lk.localParticipant)
-        }
-      }
-    }
-    lk.on(RoomEvent.ParticipantConnected, attachParticipant)
-    lk.on(RoomEvent.ParticipantDisconnected, handleParticipantDisconnected)
-    lk.on(RoomEvent.TrackMuted, muteHandler)
-    lk.on(RoomEvent.TrackUnmuted, muteHandler)
-    lk.on(RoomEvent.TrackSubscribed, handleTrackSubscribed)
-    lk.on(RoomEvent.TrackUnsubscribed, handleTrackUnsubscribed)
-    lk.on(RoomEvent.LocalTrackPublished, handleLocalTrack)
-    lk.on(RoomEvent.LocalTrackUnpublished, handleLocalTrackUnsubscribed)
     roomEl && roomEl.addEventListener('fullscreenchange', handleFullScreen)
     loading = false
   })
@@ -282,36 +68,7 @@
   let columns: number = 0
   let rows: number = 0
 
-  onDestroy(
-    infos.subscribe((data) => {
-      for (const info of data) {
-        if (info.room !== room._id) continue
-        const current = participants.find((p) => p._id === info.person)
-        if (current !== undefined) continue
-        const value: ParticipantData = {
-          _id: info.person,
-          name: info.name,
-          muted: true,
-          mirror: false,
-          connecting: true,
-          isAgent: aiPersonRef === info.person
-        }
-        participants.push(value)
-      }
-      participants = participants
-      updateStyle(getActiveParticipants(participants).length, $screenSharing)
-    })
-  )
-
   onDestroy(() => {
-    lk.off(RoomEvent.ParticipantConnected, attachParticipant)
-    lk.off(RoomEvent.ParticipantDisconnected, handleParticipantDisconnected)
-    lk.off(RoomEvent.TrackSubscribed, handleTrackSubscribed)
-    lk.off(RoomEvent.TrackUnsubscribed, handleTrackUnsubscribed)
-    lk.off(RoomEvent.LocalTrackPublished, handleLocalTrack)
-    lk.off(RoomEvent.TrackMuted, muteHandler)
-    lk.off(RoomEvent.TrackUnmuted, muteHandler)
-    lk.off(RoomEvent.LocalTrackUnpublished, handleLocalTrackUnsubscribed)
     roomEl.removeEventListener('fullscreenchange', handleFullScreen)
   })
 
@@ -375,11 +132,6 @@
 
   $: if (((document.fullscreenElement && !$isFullScreen) || $isFullScreen) && roomEl) checkFullscreen()
 
-  function getActiveParticipants (participants: ParticipantData[]): ParticipantData[] {
-    return participants.filter((p) => !p.isAgent || $infos.some(({ person }) => person === p._id))
-  }
-
-  $: activeParticipants = getActiveParticipants(participants)
 </script>
 
 <div bind:this={roomEl} class="flex-col-center w-full h-full" class:theme-dark={$isFullScreen}>
@@ -399,7 +151,7 @@
     class:mobile={$deviceInfo.isMobile}
   >
     <div class="screenContainer">
-      <video class="screen" bind:this={screen}></video>
+      <ScreenSharingView />
     </div>
     {#if withVideo}
       <div
@@ -407,9 +159,10 @@
         style={$screenSharing ? '' : gridStyle}
         class:scroll-m-0={$screenSharing}
       >
-        {#each activeParticipants as participant, i (participant._id)}
-          <ParticipantView bind:this={participantElements[i]} {...participant} />
-        {/each}
+        <ParticipantsListView room={room._id} on:participantsCount={(evt) => {
+          console.log('participantsCount', evt.detail)
+          updateStyle(evt.detail, $screenSharing)
+        }}/>
       </div>
     {/if}
   </div>
